instalei sass, react-router-dom e classesnames; implementei as routes, as páginas Map, Trucks e MainPage
