+++ teste-aiko/package.json
@@ -17,12 +17,15 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.0.1",
     "browserify": "^17.0.0",
+    "classnames": "^2.5.1",
     "leaflet": "^1.9.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.1",
     "react-leaflet-cluster": "^2.1.0",
     "react-redux": "^9.1.0",
+    "react-router-dom": "^6.26.2",
+    "sass": "^1.78.0",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
