+++ teste-aiko/src/index.css
@@ -12,13 +12,3 @@ body {
 #root {
   height: 100vh;
 }
-
-.leaflet-container {
-  height: 100%;
-}
-.cluster-icon {
-  height: 3rem;
-  width: 3rem;
-  border-radius: 50%;
-  background-color: #fff;
-}
