+++ teste-aiko/src/main.tsx
@@ -4,6 +4,7 @@ import { Provider } from "react-redux"
 import App from "./App"
 import { store } from "./app/store"
 import "./index.css"
+import Router from './routes';
 
 const container = document.getElementById("root")
 
@@ -13,7 +14,7 @@ if (container) {
   root.render(
     <React.StrictMode>
       <Provider store={store}>
-        <App />
+        <Router />
       </Provider>
     </React.StrictMode>,
   )
