+++ teste-aiko/src/pages/Map/index.css
@@ -1,0 +1,10 @@
+.leaflet-container {
+    height: 100vh;
+}
+
+.cluster-icon {
+    height: 3rem;
+    width: 3rem;
+    border-radius: 50%;
+    background-color: #fff;
+}
