+++ teste-aiko/src/pages/Trucks/Trucks.tsx
@@ -1,0 +1,7 @@
+export default function Trucks() {
+
+
+    return (
+      <div>Trucks</div>
+    )
+  }
